Add badges (#29)
Add self-hosted (well, github pages hosted) badges and reports (tests and coverage) using coverage.py and genbadge.

The reports and badges are hosted in a different repo. To add this to another repo, you'd just have to use a suitable access token as a repository secret (ie, one that has access to the qupath badges repo).