Refactor classification code to handle multiple classifications (#44)
Previously the assumption was a classification has one name.
However QuPath features can have multiple classifications:
an object with class "A: B: C" is converted to JSON something like
{[blah], "properties": {"classification": {"names": ["A", "B", "C"]}}}
whereas an object with class "A" is converted something like
{[blah], "properties": {"classification": {"name": "A"}}}

This PR removes the assumption of a single class while ensuring a single
name can be fetched using a mocked `name` property that joins on ": ",
similar to objects with multiple classes in QuPath.

Resolve #43